LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/uievents/mouse/layout_change_should_fire_mouseover-expected.txt

Test Description: Tests that the mouseover event is fired and the element has a hover effect when the element underneath the mouse cursor is changed.
Put your mouse over the red rectangle
Click the primary mouse button


FAIL Tests that the mouseover event is fired and the element has a hover effect when the element underneath the mouse cursor is changed. assert_equals: expected "mouseover,mouseenter" but got ""
